refactor: section 패키지 분리
