+++ src/main/java/com/lubycon/curriculum/section/repository/IntroSectionRepository.java
@@ -1,6 +1,6 @@
-package com.lubycon.curriculum.curriculum.repository;
+package com.lubycon.curriculum.section.repository;
 
-import com.lubycon.curriculum.curriculum.domain.IntroSection;
+import com.lubycon.curriculum.section.domain.IntroSection;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IntroSectionRepository extends JpaRepository<IntroSection, Long> {
