+++ src/main/java/com/lubycon/curriculum/section/repository/SectionRepository.java
@@ -1,6 +1,6 @@
-package com.lubycon.curriculum.curriculum.repository;
+package com.lubycon.curriculum.section.repository;
 
-import com.lubycon.curriculum.curriculum.domain.Section;
+import com.lubycon.curriculum.section.domain.Section;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SectionRepository extends JpaRepository<Section, Long> {
