+++ src/test/java/com/lubycon/curriculum/curriculum/repository/CurriculumRepositoryTest.java
@@ -4,8 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lubycon.curriculum.base.RepositoryTest;
 import com.lubycon.curriculum.curriculum.domain.Curriculum;
-import com.lubycon.curriculum.curriculum.domain.IntroSection;
-import com.lubycon.curriculum.curriculum.domain.Section;
+import com.lubycon.curriculum.section.domain.IntroSection;
+import com.lubycon.curriculum.section.domain.Section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
