+++ src/test/java/com/lubycon/curriculum/section/repository/SectionRepositoryTest.java
@@ -1,10 +1,10 @@
-package com.lubycon.curriculum.curriculum.repository;
+package com.lubycon.curriculum.section.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.lubycon.curriculum.base.RepositoryTest;
 import com.lubycon.curriculum.blog.domain.Blog;
-import com.lubycon.curriculum.curriculum.domain.Section;
+import com.lubycon.curriculum.section.domain.Section;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
